test: remove redundant test for Promise ordering
The earlier test which affirms Promise equality with subsequent calls
guarantees the behavior of sequential ordering as per the spec. It is
pointless to test this as if the test breaks it is a bug with the JS
engine and one we're unlikely able to fix (or is at least out of the
scope of this project). Our interest is only in the earlier test of
referential equality.

This test also tests the cache implementation, but this is needless as
there are other tests which affirm these semantics.